refactor endpoint status and stat
