+++ app/models/error_codes.py
@@ -13,3 +13,5 @@ class ErrorCode(str, Enum):
     PREDICTION_FAILED = "PREDICTION_FAILED"
     INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR"
     MODEL_NOT_FOUND = "MODEL_NOT_FOUND"
+    SECURITY_STATS_RETRIEVAL_ERROR = "SECURITY_STATS_RETRIEVAL_ERROR"
+    RATE_LIMIT_NOT_INITIALIZED = "RATE_LIMITER_NOT_INITIALIZED"
